remove get_fields static methods and redundant code
